#261 Fix NPE in resource close if file is innacurate. Stops creation of CassandraRunner.
